# servers/node/src/docs/openapi.yaml
openapi: "3.0.0"
info:
  version: 12.0.0
  title: Jeera Project Manager
  license:
    name: MIT
  contact:
    name: Antoine Cheron
    email: cheron.antoine@gmail.com
paths:
  /users:
    get:
      summary: List existing users public information
      operationId: listUsers
      tags:
        - users
      responses:
        '200':
          description: All existing users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailsList'
      security: 
        - bearerToken: []
    post:
      summary: Creates a new user (admin only)
      operationId: createUser
      tags:
        - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User just created
          headers:
            Location:
              description: 'URL to the resource created'
              schema:
                type: string
                format: url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetails'
          links:
            confirmEmail:
              operationId: confirmEmail
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/confirm: # to remove in v2
    get:
      summary: Confirm email ownership
      operationId: confirmEmail
      tags: 
        - users
      parameters:
        - name: token
          in: query
          description: Confirmation token received by email
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Success
          links:
            login:
              operationId: login
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user/login:
    post:
      summary: Login into the application
      operationId: login
      tags:
        - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful connection
          headers:
            Authorization:
              description: Token to use
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginConfirmation'
          links:
            logout:
              operationId: logout
            getUserDetails:
              operationId: getUserDetails
            listProjects:
              operationId: listProjects
        '403':
          description: 'Wrong credentials'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /user/logout:
    post:
      summary: Logout
      operationId: logout
      tags: 
        - users
      responses:
        '204':
          description: 'Successfully logged out - revoke token'
        '500':
          description: 'Unexpected internal server error, try again'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /user:
    get:
      summary: Get a user details
      operationId: getCurrentUserDetails
      tags:
        - users
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetails'
          links:
            listProjects:
              operationId: listProjects
            changePassword:
              operationId: updateUserPassword
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /user/{userId}:
    parameters: 
      - $ref: '#/components/parameters/userId'
    get:
      summary: Get a user details
      operationId: getUserDetails
      tags:
        - users
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetails'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /user/password:
    parameters: 
      - $ref: '#/components/parameters/userId'
    put:
      summary: Updates a user's password
      operationId: updateUserPassword
      tags:
        - users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '204':
          description: 'Password successfully updated'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
  /projects:
    get:
      summary: List the projects that the connected user can see
      operationId: listProjects
      security:
        - bearerToken: []
      tags:
        - projects
      parameters:
        - name: offset
          in: query
          description: The number of items to skip before starting to collect the result set
          required: false
          schema:
            type: integer
            format: int8
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Number of elements to get
          required: false
          schema:
            type: integer
            format: int8
            minimum: 1
            maximum: 40
            default: 5
      responses:
        '200':
          description: 'Success'
          headers:
            X-Next:
              description: Link to next page
              required: false
              schema:
                type: string
                format: url
            X-Last:
              description: Link to last page
              required: false
              schema:
                type: string
                format: url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectsList'
          links:
            createProject:
              operationId: createProject
            setTaskStatusesFlow:
              operationId: setProjectTaskStatusFlow
            setDetails:
              operationId: setProjectDetails
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Creates a project
      operationId: createProject
      tags:
        - projects
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          $ref: '#/components/responses/projectDetails'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security: 
        - bearerToken: []
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    projectDetails:
      description: Project details
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Project'
      links:
        inviteUser:
          operationId: inviteUsersToProject
          parameters:
            projectId: '$response.body#/projectId'
        createTechnicalStory:
          operationId: createTechnicalStory
          parameters:
            parentProjectId: '$response.body#/id'
        createUserStory:
          operationId: createUserStory
          parameters:
            parentProjectId: '$response.body#/id'
        listTasks:
          operationId: listTasks
          parameters:
            queryProjectId: '$response.body#/id'
        reverseArchivedState:
          operationId: switchProjectArchiveState
          parameters:
            projectId: '$response.body#/id'
        delete:
          operationId: deleteProject
          parameters:
            projectId: '$response.body#/id'
        star:
          operationId: starProject
          parameters:
            projectId: '$response.body#/id'
        analytics:
          operationId: getAnalytics
          parameters:
            resourceId: '$response.body#/id'
        setTaskStatusesFlow:
          operationId: setProjectTaskStatusFlow
          parameters:
            projectId: '$response.body#/id'
        setDetails:
          operationId: setProjectDetails
          parameters:
            projectId: '$response.body#/id'
  parameters:
    userId: 
      name: userId
      in: path
      description: Id of the user to retrieve details for
      required: true
      schema:
        type: string
        format: url
...
  schemas:
    CreateUserRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
    UserDetailsList:
      type: array
      items:
        $ref: '#/components/schemas/UserDetails'
    UserDetails:
      type: object
      properties:
        username:
          type: string
        id:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        starredProjects:
          type: array
          items:
            type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: password
      required: [ 'username', 'password' ]
    LoginConfirmation:
      type: object
      properties:
        token:
          type: string
    UpdatePasswordRequest:
      type: object
      properties:
        previousPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
      required: [ previousPassword, newPassword ]
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
      required: [ name ]
    ProjectsList:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/Project'
    Project:
      type: object
      properties:
        name:
          type: string
        id:
          type: string
        isArchived:
          type: boolean
        collaborators:
          type: array
          items:
            type: string
            description: username
        description:
          type: string
        availableTaskStatuses:
          type: array
          items:
            $ref: '#/components/schemas/TaskStatus'
        taskStatusTransitions:
          type: array
          items:
            $ref: '#/components/schemas/TaskStatusTransition'
        nextCreationStep:
          type: string
          enum: 
            - INITIALIZATION
            - TASK_FLOW_SETUP
            - INFORMATION_SETUP
            - CREATION_COMPLETED
      required: [ name, id, isArchived, collaborators ]
...
    TaskStatus:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
    TaskStatusTransition:
      type: object
      properties:
        from:
          type: string
          description: Either a comma-delimited string of TaskStatus ids or '*' to select them all
        to:
          type: string
          description: Either a comma-delimited string of TaskStatus ids or '*' to select them all
...
    UserRole:
      type: string
      enum: 
        - Product Owner
        - Developer
...
    Error:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
...